BB_psycopg2_excel2postgres/requirements.txt
psycopg2-binary==2.9.10
pandas==2.3.0
python-dotenv==1.1.1
ipykernel==6.29.5
openpyxl==3.1.5